% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInvitedSeq.R
\name{addInvitedSeq}
\alias{addInvitedSeq}
\title{A function to update barcode summaries with a selected set of DNA sequences (e.g. in case of big rearrangement in barcode sequence).}
\usage{
addInvitedSeq(pat, bc_data, dir, invited_idx = NULL, invited_bc = NULL)
}
\arguments{
\item{pat}{The barcode backbone pattern for matching.}

\item{bc_data}{The data returned by the function generateSummaries().}

\item{dir}{The directory containing the summaries.}

\item{invited_idx}{The list of invited sequences that missed index matching (A table with matching sequences and sample name. one sequence per line).}

\item{invited_bc}{The list of invited sequences that missed barcode matching (one sequence per line).}
}
\description{
A function to update barcode summaries with a selected set of DNA sequences (e.g. in case of big rearrangement in barcode sequence).
}
\details{
This function updates the summaries generated by generateSummaries() with a set of provided sequences.
}
\examples{
addInvitedSeq(pat, bc_data, getwd(), invited_idx, invited_bc)
}
\keyword{BC32}
\keyword{processing}
\keyword{update}
