% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsToCPM.R
\name{countsToCPM}
\alias{countsToCPM}
\title{A function to convert counts to counts per million (CPM).}
\usage{
countsToCPM(counts)
}
\arguments{
\item{counts}{A merged count summary generated by mergeSummaries().}
}
\description{
A function to convert counts to counts per million (CPM).
}
\details{
This function returns a table of counts converted to counts per million with an additional column
containing the sum of CPMs across samples.
}
\examples{
cpm <- countsToCPM(counts)
}
\keyword{BC32}
\keyword{CPM}
\keyword{convert}
\keyword{processing}
