% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsToFreq.R
\name{countsToFreq}
\alias{countsToFreq}
\title{A function to convert counts to frequencies.}
\usage{
countsToFreq(counts)
}
\arguments{
\item{counts}{A merged count summary generated by mergeSummaries().}
}
\description{
A function to convert counts to frequencies.
}
\details{
This function returns a table of counts converted to frequencies with an additional column
containing the sum of frequencies across samples.
}
\examples{
freq <- countsToFreq(counts)
}
\keyword{BC32}
\keyword{convert}
\keyword{frequencies}
\keyword{processing}
