% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreBC.R
\name{exploreBC}
\alias{exploreBC}
\title{A function to explore a processed BarSeq run.}
\usage{
exploreBC(dir)
}
\arguments{
\item{dir}{The directory containing the summaries.}
}
\description{
A function to explore a processed BarSeq run.
}
\details{
This function starts a shiny app to explore the results of a BarSeq run.
}
\examples{
exploreBC(getwd())
}
\keyword{BC32}
\keyword{explore}
\keyword{results}
