% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummaries.R
\name{generateSummaries}
\alias{generateSummaries}
\title{A function to generate summaries from BC32 fastq files.}
\usage{
generateSummaries(
  pat,
  restriction,
  sampname,
  base_q = 20,
  idx_mis = 1,
  bb_mis = 1,
  indels = 1
)
}
\arguments{
\item{pat}{The barcode backbone pattern for matching.}

\item{restriction}{The nucleotide sequence of the restriction site flanking the barcode at the 3' end.}

\item{sampname}{A data frame with 3 columns containing the name of each sample to
process, its file name and the expected multiplex index sequence.}

\item{base_q}{The minimum mean base quality in the first 90 nucleotides required to keep
a read for the processing. Defaults to 20.}

\item{idx_mis}{The number of mismatches allowed during index matching. Defaults to 1.}

\item{bb_mis}{The number of mismatches allowed during barcode matching. Defaults to 1.}

\item{indels}{The number of indels allowed during barcode matching. This strongly impacts
the speed of the function. Try to keep it low. Defaults to 1.}
}
\value{
Returns data in a nested list of samples containing a list of sequences not matching the index,
QC data and sequences not matching the barcode index.
}
\description{
A function to generate summaries from BC32 fastq files.
}
\details{
This function extracts the number of different barcode sequences found in a provided
list of fastq files and export a summary table, quality control plots, and a list of the most
common sequences that failed index or barcode matching.
}
\examples{
bc_data <- generateSummaries(pat = "CTAGCCAGTT",
                             restriction = "CTCGAG"
                             sampname = sampname)
}
\keyword{BC32}
\keyword{QC}
\keyword{processing}
