% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBC.R
\name{initBC}
\alias{initBC}
\title{Initialization function (loading of the required packages).}
\usage{
initBC()
}
\description{
Initialization function (loading of the required packages).
}
\details{
This function loads the required packages
}
\examples{
initBC()
}
\keyword{BC32}
\keyword{initialize}
