% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSummaries.R
\name{mergeSummaries}
\alias{mergeSummaries}
\title{A function to merge multiple count summaries into one table.}
\usage{
mergeSummaries(sampname, dir, min.count = 10, cleancol = F)
}
\arguments{
\item{sampname}{A data frame with 3 columns containing the name of each sample to
process, its file name and the expected multiplex index sequence.}

\item{dir}{The directory containing the summaries.}

\item{min.count}{Counts below this number are replaced by zero (considered as noise). Defaults to 10.}

\item{cleancol}{If true, columns that contain only zeros will be removed. Defaults to F.}
}
\description{
A function to merge multiple count summaries into one table.
}
\details{
This function merge all count summaries in the sampname table together in one data frame. Counts below
min.count are replaced by zeros.
}
\examples{
counts <- mergeSummaries(sampname,
                         getwd(),
                         10,
                         F)
}
\keyword{BC32}
\keyword{merge}
\keyword{processing}
\keyword{update}
