% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolBC.R
\name{poolBC}
\alias{poolBC}
\title{A function to pool similar barcodes according to Damerau-Levenshtein distance.}
\usage{
poolBC(sampname, dir, dl = 12)
}
\arguments{
\item{sampname}{A data frame with 3 columns containing the name of each sample to
process, its file name and the expected multiplex index sequence.}

\item{dir}{The directory containing the summaries.}

\item{dl}{The maximum Damerau-Levenshtein distance to consider barcodes as identical and pool them together. Defaults to 12.}
}
\description{
A function to pool similar barcodes according to Damerau-Levenshtein distance.
}
\details{
This function updates the count summaries by pooling together similar barcodes (according to the given Damerau-Levenshtein distance).
}
\examples{
poolBC(sampname, getwd(), 12)
}
\keyword{BC32}
\keyword{Damerau-Levenshtein}
\keyword{pool}
\keyword{processing}
\keyword{update}
