% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmatchPattern2.R
\name{vmatchPattern2}
\alias{vmatchPattern2}
\title{String searching function.}
\usage{
vmatchPattern2(
  pattern,
  subject,
  max.mismatch = 0,
  min.mismatch = 0,
  with.indels = FALSE,
  fixed = TRUE,
  algorithm = "auto"
)
}
\arguments{
\item{pattern}{The pattern string.}

\item{subject}{The String object for matching.}

\item{max.mismatch, min.mismatch}{The minimum and maximum number of mismatching letters allowed.}

\item{with.indels}{If TRUE then indels are allowed.}

\item{fixed}{If TRUE (the default), an IUPAC ambiguity code in the pattern can only match
the same code in the subject, and vice versa.}

\item{algorithm}{One of the following: "auto", "naive-exact", "naive-inexact", "boyer-moore",
"shift-or" or "indels".}

\item{...}{Additional arguments for methods.}
}
\description{
A functions for finding all the occurrences (aka "matches" or "hits") of a given pattern (typically
short) in a (typically long) reference sequence or set of reference sequences (aka the subject).
This is an updated version of vmatchPattern that take indels into account.
Provided by Hervé Pagès on https://support.bioconductor.org/p/58350/
}
\details{
See vmatchPattern for details.
}
\examples{
x <- DNAString("AAGCGCGATATG")
m1 <- matchPattern("GCNNNAT", x)
m1
}
\keyword{matching}
\keyword{pattern}
